PackageManager update
