+++ .gitignore
@@ -1,23 +1,3 @@
-# Xcode
 .DS_Store
-build/
-*.pbxuser
-!default.pbxuser
-*.mode1v3
-!default.mode1v3
-*.mode2v3
-!default.mode2v3
-*.perspectivev3
-!default.perspectivev3
-xcuserdata
-*.xccheckout
-*.moved-aside
-DerivedData
-*.hmap
-*.ipa
-*.xcuserstate
-*.xcscmblueprint
-*.resolved
-Carthage/Build
-Carthage/Checkouts
-
+.build/
+.swiftpm/
