--- SwiftMD5/SwiftMD5.h
@@ -1,11 +1,0 @@
-#import <Cocoa/Cocoa.h>
-
-//! Project version number for SwiftMD5.
-FOUNDATION_EXPORT double SwiftMD5VersionNumber;
-
-//! Project version string for SwiftMD5.
-FOUNDATION_EXPORT const unsigned char SwiftMD5VersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <SwiftMD5/PublicHeader.h>
-
-
